tests/wpt/tests/fetch/metadata/generated/element-input-image.https.sub.html

<!DOCTYPE html>
<!--
This test was procedurally generated. Please do not modify it directly.
Sources:
- fetch/metadata/tools/fetch-metadata.conf.yml
- fetch/metadata/tools/templates/element-input-image.sub.html
-->
<html lang="en">
  <meta charset="utf-8">
  <title>HTTP headers on request for HTML "input" element with type="button"</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/fetch/metadata/resources/helper.sub.js"></script>
  <body>
  <script>
  'use strict';

  function induceRequest(url, test) {
    const input = document.createElement('input');
    input.setAttribute('type', 'image');

    document.body.appendChild(input);
    test.add_cleanup(() => input.remove());

    return new Promise((resolve) => {
        input.onload = input.onerror = resolve;
        input.setAttribute('src', url);
      });
  }

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(makeRequestURL(key, ['httpsOrigin']), t)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_array_equals(headers['sec-fetch-site'], ['same-origin']);
        });
  }, 'sec-fetch-site - Same origin, no attributes');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(makeRequestURL(key, ['httpsCrossSite']), t)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_array_equals(headers['sec-fetch-site'], ['cross-site']);
        });
  }, 'sec-fetch-site - Cross-site, no attributes');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(makeRequestURL(key, ['httpsSameSite']), t)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_array_equals(headers['sec-fetch-site'], ['same-site']);
        });
  }, 'sec-fetch-site - Same site, no attributes');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(makeRequestURL(key, ['httpsOrigin', 'httpsCrossSite', 'httpsOrigin']), t)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_array_equals(headers['sec-fetch-site'], ['cross-site']);
        });
  }, 'sec-fetch-site - Same-Origin -> Cross-Site -> Same-Origin redirect, no attributes');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(makeRequestURL(key, ['httpsOrigin', 'httpsSameSite', 'httpsOrigin']), t)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_array_equals(headers['sec-fetch-site'], ['same-site']);
        });
  }, 'sec-fetch-site - Same-Origin -> Same-Site -> Same-Origin redirect, no attributes');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(makeRequestURL(key, ['httpsCrossSite', 'httpsOrigin']), t)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_array_equals(headers['sec-fetch-site'], ['cross-site']);
        });
  }, 'sec-fetch-site - Cross-Site -> Same Origin, no attributes');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(makeRequestURL(key, ['httpsCrossSite', 'httpsSameSite']), t)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_array_equals(headers['sec-fetch-site'], ['cross-site']);
        });
  }, 'sec-fetch-site - Cross-Site -> Same-Site, no attributes');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(makeRequestURL(key, ['httpsCrossSite', 'httpsCrossSite']), t)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_array_equals(headers['sec-fetch-site'], ['cross-site']);
        });
  }, 'sec-fetch-site - Cross-Site -> Cross-Site, no attributes');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(makeRequestURL(key, ['httpsOrigin', 'httpsOrigin']), t)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_array_equals(headers['sec-fetch-site'], ['same-origin']);
        });
  }, 'sec-fetch-site - Same-Origin -> Same Origin, no attributes');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(makeRequestURL(key, ['httpsOrigin', 'httpsSameSite']), t)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_array_equals(headers['sec-fetch-site'], ['same-site']);
        });
  }, 'sec-fetch-site - Same-Origin -> Same-Site, no attributes');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(makeRequestURL(key, ['httpsOrigin', 'httpsCrossSite']), t)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_array_equals(headers['sec-fetch-site'], ['cross-site']);
        });
  }, 'sec-fetch-site - Same-Origin -> Cross-Site, no attributes');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(makeRequestURL(key, ['httpsSameSite', 'httpsOrigin']), t)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_array_equals(headers['sec-fetch-site'], ['same-site']);
        });
  }, 'sec-fetch-site - Same-Site -> Same Origin, no attributes');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(makeRequestURL(key, ['httpsSameSite', 'httpsSameSite']), t)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_array_equals(headers['sec-fetch-site'], ['same-site']);
        });
  }, 'sec-fetch-site - Same-Site -> Same-Site, no attributes');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(makeRequestURL(key, ['httpsSameSite', 'httpsCrossSite']), t)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_array_equals(headers['sec-fetch-site'], ['cross-site']);
        });
  }, 'sec-fetch-site - Same-Site -> Cross-Site, no attributes');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(makeRequestURL(key, ['httpsOrigin', 'httpOrigin', 'httpsOrigin']), t)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_array_equals(headers['sec-fetch-site'], ['cross-site']);
        });
  }, 'sec-fetch-site - HTTPS downgrade-upgrade, no attributes');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(makeRequestURL(key, []), t)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-mode');
          assert_array_equals(headers['sec-fetch-mode'], ['no-cors']);
        });
  }, 'sec-fetch-mode - no attributes');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(makeRequestURL(key, []), t)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-dest');
          assert_array_equals(headers['sec-fetch-dest'], ['image']);
        });
  }, 'sec-fetch-dest - no attributes');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(makeRequestURL(key, []), t)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_not_own_property(headers, 'sec-fetch-user');
        });
  }, 'sec-fetch-user - no attributes');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(makeRequestURL(key, ['httpsCrossSite']), t)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-storage-access');
          assert_array_equals(headers['sec-fetch-storage-access'], ['none']);
        });
  }, 'sec-fetch-storage-access - Cross-site, no attributes');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(makeRequestURL(key, ['httpsSameSite']), t)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_not_own_property(headers, 'sec-fetch-storage-access');
        });
  }, 'sec-fetch-storage-access - Same site, no attributes');
  </script>
  </body>
</html>
